Change value of _ENTERED from 2 to 0
This change allows the solidity compiler to optimize the status check. Essentially, checking for (in)equality with a non-zero number requires a `push` and `eq` opcode, whereas checking for (in)equality with zero can make use of the `iszero` opcode instead.

Importantly, the cost and refund amounts for the storage operations remain the same. See https://eips.ethereum.org/EIPS/eip-3529 for details, but essentially going 1 -> 2 -> 1 costs the same and gives the same refund as going 1 -> 0 -> 1.

Lastly, the `sstore(0, 0)` operation is also slightly optimized as it can be done using a `push` and a `dup` instead of two `push` opcodes.

In total, this reduces the bytecode size by 4 bytes and gas costs by 6 gas.